Move engine structs into `models`
This lets us reuse them in a more sensible way
